+++ kauldron/modules/__init__.py
@@ -32,6 +32,7 @@ from kauldron.modules.knn_types import TransformerBlock
 
 # Modules
 from kauldron.modules.adapter import ExternalModule
+from kauldron.modules.adapter import WrapperModule
 from kauldron.modules.misc import Dropout
 from kauldron.modules.misc import DummyModel
 from kauldron.modules.misc import Identity